backend/requirements.txt
# Core
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
gunicorn==21.2.0

# XML / Fiscal
signxml==4.3.1
lxml==6.0.2
cryptography==46.0.5

# Media
Pillow==10.2.0

# CORS
django-cors-headers==4.3.1

# API Documentation
drf-spectacular==0.27.2

# Background Tasks
celery[redis]==5.4.0
django-celery-beat==2.6.0
redis==5.0.8

# AI
anthropic>=0.39.0

# HTTP
requests>=2.31.0

# Monitoring
sentry-sdk[django,celery,redis]>=2.0.0

# Exports
openpyxl==3.1.5
reportlab==4.2.5

# Audit
django-auditlog==3.0.0

# Security
argon2-cffi>=23.1.0
pyotp==2.9.0
qrcode[pil]==7.4.2
django-ratelimit==4.1.0
django-csp==3.8
bleach==6.1.0
django-dbbackup==4.1.0

# Testing
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==6.0.0
factory-boy==3.3.1
pytest-asyncio==0.24.0

# Channels (WebSockets)
channels[daphne]==4.1.0
channels-redis==4.2.0
